requirements.txt
Django==5.0.3
djangorestframework==3.15.0
psycopg2==2.9.2
psycopg2-binary==2.9.9
pytest==8.1.1